catalog_parser.py
import requests
from bs4 import BeautifulSoup

def parse_catalog():
    url = "https://www.teashop.by/catalog/"
    response = requests.get(url)
    soup = BeautifulSoup(response.text, 'html.parser')
    # TODO: Implement actual parsing logic
    teas = []  # List of dicts with tea data
    # Example: teas.append({'name': ..., 'category': ..., ...})
    return teas

def update_database(teas):
    # TODO: Implement database update logic
    pass 